document-service/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Pillow==10.2.0
anthropic==0.9.0
pdf2image==1.16.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.114.0
requests==2.31.0